webapi/webapi-resourcetiming-w3c-tests/resourcetiming/resource-timing-workerstart.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<meta name="timeout" content="long">
<title>PerformanceResourceTiming.workerStart</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://w3c.github.io/resource-timing/#performanceresourcetiming">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<div id="log"></div>
<script>

test(function() {
  var resources = window.performance.getEntriesByType("resource");
  for(var i=0; i < resources.length; i++) {
    assert_equals(resources[i].workerStart, 0, "resources[i].workerStart");
  }
}, "Check that workerStart should return zero if there is no active worker.");

promise_test(function(t) {
  var scope = "resources/blank-iframe.html"
  var scriptURL = "resources/service-worker.js";
  return service_worker_unregister_and_register(t, scriptURL, scope)
    .then(function(registration) {
      return wait_for_state(t, registration.installing, "activated");
    })
    .then(function(state) {
      return with_iframe(scope);
    })
    .then(function(frame) {
      var resoucreURL = normalizeURL(scope);
      var entry = window.performance.getEntriesByName(resoucreURL)[0];
      assert_greater_than(entry.workerStart, 0);
      assert_greater_than_equal(entry.workerStart, entry.startTime);
      assert_less_than_equal(entry.workerStart, entry.fetchStart);
      frame.remove();
      return service_worker_unregister_and_done(t);
    });
}, "Check that workerStart's value is greater than zero, but less than fetchStart" +
   "when the current browsing have an active service worker");

</script>
